Refactor: introduce Tracking namespace + rename RCAdMob → Tracking.Adapter
First step in regrouping the AdMob adapter under a single namespace.

Internal-only refactor — no public API changes:
* New `Tracking` umbrella enum (caseless) holds `Adapter`, `Tracker`,
  `PurchasesTracker`.
* `RCAdMob` class → `Tracking.Adapter`; its file is split into
  `Tracking/{Tracking,Adapter,Tracker,PurchasesTracker}.swift`.
* `AdTracking` protocol → `Tracking.Tracker`.
* `PurchasesAdTracker` → `Tracking.PurchasesTracker`.
* `rcAdMob:` parameter / `rcAdMob` property → `adapter:` / `adapter`
  across the four delegate wrappers, the GMA `loadAndTrack` extensions,
  and tests.

The four `RCAdMob*` delegate wrapper class names, `RCFullScreenAdTracking`,
and the `Tracking/` file moves of the wrappers themselves are deferred to
commit 2 for a smaller diff per step.
