+++ AdapterSDKs/RevenueCatAdMob/Tests/RevenueCatAdMobTests/RCAdMobTrackingTests.swift
@@ -7,7 +7,7 @@ import GoogleMobileAds
 @_spi(Experimental) @testable import RevenueCatAdMob
 
 @available(iOS 15.0, *)
-final class MockAdTracker: AdTracking {
+final class MockAdTracker: Tracking.Tracker {
     var isConfigured: Bool = true
 
     struct Call: Equatable {
@@ -76,7 +76,7 @@ final class MockAdTracker: AdTracking {
     }
 }
 
-// MARK: - Core RCAdMob tracking tests
+// MARK: - Core Tracking.Adapter tracking tests
 
 @MainActor
 @available(iOS 15.0, *)
@@ -84,16 +84,16 @@ final class MockAdTracker: AdTracking {
 final class RCAdMobTrackingTests: RCAdMobTestCase {
 
     private var mockTracker: MockAdTracker!
-    private var rcAdMob: RCAdMob!
+    private var adapter: Tracking.Adapter!
 
     override func setUp() {
         super.setUp()
         self.mockTracker = MockAdTracker()
-        self.rcAdMob = RCAdMob(tracker: self.mockTracker)
+        self.adapter = Tracking.Adapter(tracker: self.mockTracker)
     }
 
     func testTrackLoadedCallsTrackerWithCorrectData() {
-        self.rcAdMob.trackLoaded(
+        self.adapter.trackLoaded(
             responseInfo: nil,
             placement: "home_banner",
             adUnitID: "ca-app-pub-123",
@@ -111,7 +111,7 @@ final class RCAdMobTrackingTests: RCAdMobTestCase {
 
     func testTrackFailedToLoadCallsTrackerWithCorrectData() {
         let error = NSError(domain: "com.google.ads", code: 3, userInfo: nil)
-        self.rcAdMob.trackFailedToLoad(
+        self.adapter.trackFailedToLoad(
             placement: "feed",
             adUnitID: "ca-app-pub-456",
             adFormat: .banner,
@@ -128,7 +128,7 @@ final class RCAdMobTrackingTests: RCAdMobTestCase {
     }
 
     func testTrackDisplayedCallsTrackerWithCorrectData() {
-        self.rcAdMob.trackDisplayed(
+        self.adapter.trackDisplayed(
             responseInfo: nil,
             placement: "screen_top",
             adUnitID: "ca-app-pub-789",
@@ -145,7 +145,7 @@ final class RCAdMobTrackingTests: RCAdMobTestCase {
     }
 
     func testTrackOpenedCallsTrackerWithCorrectData() {
-        self.rcAdMob.trackOpened(
+        self.adapter.trackOpened(
             responseInfo: nil,
             placement: nil,
             adUnitID: "ca-app-pub-000",
@@ -163,7 +163,7 @@ final class RCAdMobTrackingTests: RCAdMobTestCase {
 
     func testTrackingNoOpsWhenNotConfigured() {
         self.mockTracker.isConfigured = false
-        self.rcAdMob.trackLoaded(
+        self.adapter.trackLoaded(
             responseInfo: nil,
             placement: "test",
             adUnitID: "unit",
@@ -184,7 +184,7 @@ final class RCAdMobTrackingTests: RCAdMobTestCase {
         )
 
         do {
-            let _: FakeFullScreenAd = try await self.rcAdMob.handleLoadOutcome(
+            let _: FakeFullScreenAd = try await self.adapter.handleLoadOutcome(
                 loadAd: { throw error },
                 context: context
             )
@@ -212,7 +212,7 @@ final class RCAdMobTrackingTests: RCAdMobTestCase {
             paidEventHandler: nil
         )
 
-        let result = try await self.rcAdMob.handleLoadOutcome(
+        let result = try await self.adapter.handleLoadOutcome(
             loadAd: { fakeAd },
             context: context
         )
@@ -234,7 +234,7 @@ final class RCAdMobTrackingTests: RCAdMobTestCase {
             fullScreenContentDelegate: nil, paidEventHandler: nil
         )
 
-        _ = try await self.rcAdMob.handleLoadOutcome(
+        _ = try await self.adapter.handleLoadOutcome(
             loadAd: { fakeAd }, context: context
         )
 
@@ -249,7 +249,7 @@ final class RCAdMobTrackingTests: RCAdMobTestCase {
             fullScreenContentDelegate: nil, paidEventHandler: nil
         )
 
-        _ = try await self.rcAdMob.handleLoadOutcome(
+        _ = try await self.adapter.handleLoadOutcome(
             loadAd: { fakeAd }, context: context
         )
 
@@ -257,7 +257,7 @@ final class RCAdMobTrackingTests: RCAdMobTestCase {
     }
 
     func testTrackRevenueCallsTrackerWithCorrectRevenueData() {
-        self.rcAdMob.trackRevenue(
+        self.adapter.trackRevenue(
             placement: "reward_screen",
             adUnitID: "ca-app-pub-reward",
             adFormat: .rewarded,
@@ -281,7 +281,7 @@ final class RCAdMobTrackingTests: RCAdMobTestCase {
     func testTrackRevenueNoOpsWhenNotConfigured() {
         self.mockTracker.isConfigured = false
 
-        self.rcAdMob.trackRevenue(
+        self.adapter.trackRevenue(
             placement: "test",
             adUnitID: "unit",
             adFormat: .interstitial,
@@ -302,7 +302,7 @@ final class RCAdMobTrackingTests: RCAdMobTestCase {
             paidEventHandler: nil
         )
 
-        _ = try await self.rcAdMob.handleLoadOutcome(
+        _ = try await self.adapter.handleLoadOutcome(
             loadAd: { fakeAd }, context: context
         )
 
@@ -329,7 +329,7 @@ final class RCAdMobTrackingTests: RCAdMobTestCase {
             paidEventHandler: { _ in userHandlerCalled = true }
         )
 
-        _ = try await self.rcAdMob.handleLoadOutcome(
+        _ = try await self.adapter.handleLoadOutcome(
             loadAd: { fakeAd }, context: context
         )
 
@@ -341,7 +341,7 @@ final class RCAdMobTrackingTests: RCAdMobTestCase {
 
     func testTrackFailedToLoadCapturesErrorCode() {
         let error = NSError(domain: "com.google.ads", code: 42, userInfo: nil)
-        self.rcAdMob.trackFailedToLoad(
+        self.adapter.trackFailedToLoad(
             placement: "feed",
             adUnitID: "ca-app-pub-456",
             adFormat: .banner,
@@ -353,7 +353,7 @@ final class RCAdMobTrackingTests: RCAdMobTestCase {
     }
 
     func testTrackLoadedExtractsResponseInfoFields() {
-        self.rcAdMob.trackLoaded(
+        self.adapter.trackLoaded(
             responseInfo: Self.makeResponseInfoPlaceholder(),
             placement: "home",
             adUnitID: "ca-app-pub-123",
@@ -367,7 +367,7 @@ final class RCAdMobTrackingTests: RCAdMobTestCase {
     }
 
     func testTrackRevenueExtractsResponseInfoFields() {
-        self.rcAdMob.trackRevenue(
+        self.adapter.trackRevenue(
             placement: "home",
             adUnitID: "ca-app-pub-123",
             adFormat: .interstitial,
@@ -392,7 +392,7 @@ final class RCAdMobTrackingTests: RCAdMobTestCase {
             paidEventHandler: nil
         )
 
-        _ = try await self.rcAdMob.handleLoadOutcome(
+        _ = try await self.adapter.handleLoadOutcome(
             loadAd: { fakeAd }, context: context
         )
 
@@ -411,7 +411,7 @@ final class RCAdMobTrackingTests: RCAdMobTestCase {
     }
 
     func testNilAdUnitIDMapsToEmptyString() {
-        self.rcAdMob.trackLoaded(
+        self.adapter.trackLoaded(
             responseInfo: nil,
             placement: "test",
             adUnitID: nil,
@@ -434,7 +434,7 @@ final class RCAdMobTrackingTests: RCAdMobTestCase {
             paidEventHandler: nil
         )
 
-        _ = try await self.rcAdMob.handleLoadOutcome(
+        _ = try await self.adapter.handleLoadOutcome(
             loadAd: { fakeAd }, context: context
         )
         await MainActor.run {
@@ -455,10 +455,10 @@ final class RCAdMobTrackingTests: RCAdMobTestCase {
             fullScreenContentDelegate: nil, paidEventHandler: nil
         )
 
-        _ = try await self.rcAdMob.handleLoadOutcome(loadAd: { fakeAd }, context: context)
+        _ = try await self.adapter.handleLoadOutcome(loadAd: { fakeAd }, context: context)
 
         let spy = FullScreenContentDelegateSpy()
-        self.rcAdMob.updateFullScreenContentDelegate(on: fakeAd, newDelegate: spy)
+        self.adapter.updateFullScreenContentDelegate(on: fakeAd, newDelegate: spy)
 
         let presentingAd = FakeFullScreenPresentingAd()
         fakeAd.fullScreenContentDelegate?.adDidRecordImpression?(presentingAd)
@@ -474,13 +474,13 @@ final class RCAdMobTrackingTests: RCAdMobTestCase {
             fullScreenContentDelegate: nil, paidEventHandler: nil
         )
 
-        _ = try await self.rcAdMob.handleLoadOutcome(loadAd: { fakeAd }, context: context)
+        _ = try await self.adapter.handleLoadOutcome(loadAd: { fakeAd }, context: context)
 
         // Simulate developer overwriting the delegate directly
         fakeAd.fullScreenContentDelegate = FullScreenContentDelegateSpy()
 
         let spy = FullScreenContentDelegateSpy()
-        self.rcAdMob.updateFullScreenContentDelegate(on: fakeAd, newDelegate: spy)
+        self.adapter.updateFullScreenContentDelegate(on: fakeAd, newDelegate: spy)
 
         let presentingAd = FakeFullScreenPresentingAd()
         fakeAd.fullScreenContentDelegate?.adDidRecordImpression?(presentingAd)
@@ -498,9 +498,9 @@ final class RCAdMobTrackingTests: RCAdMobTestCase {
             fullScreenContentDelegate: initialSpy, paidEventHandler: nil
         )
 
-        _ = try await self.rcAdMob.handleLoadOutcome(loadAd: { fakeAd }, context: context)
+        _ = try await self.adapter.handleLoadOutcome(loadAd: { fakeAd }, context: context)
 
-        self.rcAdMob.updateFullScreenContentDelegate(on: fakeAd, newDelegate: nil)
+        self.adapter.updateFullScreenContentDelegate(on: fakeAd, newDelegate: nil)
 
         let presentingAd = FakeFullScreenPresentingAd()
         fakeAd.fullScreenContentDelegate?.adDidRecordImpression?(presentingAd)
@@ -513,7 +513,7 @@ final class RCAdMobTrackingTests: RCAdMobTestCase {
         let fakeAd = FakeFullScreenAd()
         let spy = FullScreenContentDelegateSpy()
 
-        self.rcAdMob.updateFullScreenContentDelegate(on: fakeAd, newDelegate: spy)
+        self.adapter.updateFullScreenContentDelegate(on: fakeAd, newDelegate: spy)
 
         XCTAssertTrue(fakeAd.fullScreenContentDelegate === spy)
     }
@@ -525,10 +525,10 @@ final class RCAdMobTrackingTests: RCAdMobTestCase {
             fullScreenContentDelegate: nil, paidEventHandler: nil
         )
 
-        _ = try await self.rcAdMob.handleLoadOutcome(loadAd: { fakeAd }, context: context)
+        _ = try await self.adapter.handleLoadOutcome(loadAd: { fakeAd }, context: context)
 
         let spy = FullScreenContentDelegateSpy()
-        self.rcAdMob.updateFullScreenContentDelegate(on: fakeAd, newDelegate: spy)
+        self.adapter.updateFullScreenContentDelegate(on: fakeAd, newDelegate: spy)
 
         fakeAd.paidEventHandler?(Self.makeAdValuePlaceholder())
 
@@ -560,17 +560,17 @@ final class RCAdMobTrackingTests: RCAdMobTestCase {
 final class RCAdMobDelegateTrackingTests: RCAdMobTestCase {
 
     private var mockTracker: MockAdTracker!
-    private var rcAdMob: RCAdMob!
+    private var adapter: Tracking.Adapter!
 
     override func setUp() {
         super.setUp()
         self.mockTracker = MockAdTracker()
-        self.rcAdMob = RCAdMob(tracker: self.mockTracker)
+        self.adapter = Tracking.Adapter(tracker: self.mockTracker)
     }
 
     func testBannerDelegateTracksLoadedViaTracker() {
         let bannerDelegate = RCAdMobBannerViewDelegate(
-            rcAdMob: self.rcAdMob, delegate: nil, placement: "banner_top"
+            adapter: self.adapter, delegate: nil, placement: "banner_top"
         )
         let fakeBanner = GoogleMobileAds.BannerView()
         fakeBanner.adUnitID = "ca-app-pub-banner"
@@ -584,7 +584,7 @@ final class RCAdMobDelegateTrackingTests: RCAdMobTestCase {
 
     func testBannerDelegateTracksFailedToLoadViaTracker() {
         let bannerDelegate = RCAdMobBannerViewDelegate(
-            rcAdMob: self.rcAdMob, delegate: nil, placement: "banner_bottom"
+            adapter: self.adapter, delegate: nil, placement: "banner_bottom"
         )
         let fakeBanner = GoogleMobileAds.BannerView()
         fakeBanner.adUnitID = "ca-app-pub-banner"
@@ -598,7 +598,7 @@ final class RCAdMobDelegateTrackingTests: RCAdMobTestCase {
 
     func testBannerDelegateTracksImpressionViaTracker() {
         let bannerDelegate = RCAdMobBannerViewDelegate(
-            rcAdMob: self.rcAdMob, delegate: nil, placement: "banner_mid"
+            adapter: self.adapter, delegate: nil, placement: "banner_mid"
         )
         let fakeBanner = GoogleMobileAds.BannerView()
         fakeBanner.adUnitID = "ca-app-pub-banner"
@@ -611,7 +611,7 @@ final class RCAdMobDelegateTrackingTests: RCAdMobTestCase {
 
     func testBannerDelegateTracksClickViaTracker() {
         let bannerDelegate = RCAdMobBannerViewDelegate(
-            rcAdMob: self.rcAdMob, delegate: nil, placement: "banner_mid"
+            adapter: self.adapter, delegate: nil, placement: "banner_mid"
         )
         let fakeBanner = GoogleMobileAds.BannerView()
         fakeBanner.adUnitID = "ca-app-pub-banner"
@@ -625,7 +625,7 @@ final class RCAdMobDelegateTrackingTests: RCAdMobTestCase {
     func testFullScreenDelegateTracksImpressionViaTracker() {
         let fakeAd = FakeFullScreenPresentingAd()
         let delegate = RCAdMobFullScreenContentDelegate(
-            rcAdMob: self.rcAdMob,
+            adapter: self.adapter,
             delegate: nil,
             placement: "interstitial_mid",
             adUnitID: "ca-app-pub-interstitial",
@@ -643,7 +643,7 @@ final class RCAdMobDelegateTrackingTests: RCAdMobTestCase {
     func testFullScreenDelegateTracksClickViaTracker() {
         let fakeAd = FakeFullScreenPresentingAd()
         let delegate = RCAdMobFullScreenContentDelegate(
-            rcAdMob: self.rcAdMob,
+            adapter: self.adapter,
             delegate: nil,
             placement: "interstitial_mid",
             adUnitID: "ca-app-pub-interstitial",
@@ -659,7 +659,7 @@ final class RCAdMobDelegateTrackingTests: RCAdMobTestCase {
 
     func testNativeDelegateTracksImpressionViaTracker() {
         let nativeDelegate = RCAdMobNativeAdDelegate(
-            rcAdMob: self.rcAdMob, delegate: nil, placement: "native_feed", adUnitID: "ca-app-pub-native"
+            adapter: self.adapter, delegate: nil, placement: "native_feed", adUnitID: "ca-app-pub-native"
         )
         let fakeNativeAd = GoogleMobileAds.NativeAd()
 
@@ -672,7 +672,7 @@ final class RCAdMobDelegateTrackingTests: RCAdMobTestCase {
 
     func testNativeDelegateTracksClickViaTracker() {
         let nativeDelegate = RCAdMobNativeAdDelegate(
-            rcAdMob: self.rcAdMob, delegate: nil, placement: "native_feed", adUnitID: "ca-app-pub-native"
+            adapter: self.adapter, delegate: nil, placement: "native_feed", adUnitID: "ca-app-pub-native"
         )
         let fakeNativeAd = GoogleMobileAds.NativeAd()
